Move `ToolbarGroup` inside `toolbar-group` subfolder
